separated outline abstraction from implementation
