--- src/main/java/entities/course/outline/OutlineFactory.java
@@ -1,7 +1,0 @@
-package entities.course.outline;
-
-public class OutlineFactory implements OutlineFactoryInterface{
-    public Outline create() {
-        return new Outline();
-    }
-}
