--- src/main/java/entities/course/outline/OutlineFactoryInterface.java
@@ -1,5 +1,0 @@
-package entities.course.outline;
-
-public interface OutlineFactoryInterface {
-    Outline create();
-}
