+++ src/main/java/inMemoryDB/entities/OutlineImpl.java
@@ -1,0 +1,41 @@
+package inMemoryDB.entities;
+
+import entities.assessment.Assessment;
+import entities.course.outline.Outline;
+
+import java.util.ArrayList;
+
+public class OutlineImpl extends Outline {
+
+    private ArrayList<Assessment> assessments;
+
+    public OutlineImpl() {
+        this.assessments = new ArrayList<>();
+    }
+
+    @Override
+    public ArrayList<Assessment> getAssessments() {
+        return assessments;
+    }
+
+    @Override
+    public void addAssessment(Assessment assessment) {
+        if (!assessments.contains(assessment)) {
+            assessments.add(assessment);
+        }
+    }
+
+    @Override
+    public void removeAssessment(Assessment assessment) {
+        assessments.remove(assessment);
+    }
+
+    @Override
+    public Assessment getAssessment(int index) throws IndexOutOfBoundsException {
+        try {
+            return assessments.get(index);
+        } catch (IndexOutOfBoundsException e) {
+            throw new IndexOutOfBoundsException("Index out of bounds");
+        }
+    }
+}
